requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
stripe==7.9.0
Pillow==10.1.0
python-dotenv==1.0.0
reportlab==4.0.7
Werkzeug==3.0.1
email-validator==2.1.0
requests==2.31.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
